✨ feat(extensions): convert files to directory layout with tests
